have `sprite.{x,y}` be `uintptr_t` as well
this way there's no need to specify the extra position arguments in `mvspr()` and the enemy movement 'actually' works

+++ sprites.h
@@ -4,11 +4,11 @@
 
 // a sprite here has a
 typedef struct SPRITE {
-	unsigned int y, x; // starting position
-	unsigned int h, w; // height, width
+	uintptr_t    y, x; // starting and current player position
+	unsigned int h, w; // height & width
 	unsigned int hp; // health
 	unsigned int hit[2]; // hitbox (0: y, 1: x)
-	const char* skin; // and how it looks like
+	const char* skin; // how it looks like
 	WINDOW* win; // and where it's from
 	WINDOW* hud; // and its own HUD
 } SPRITE;
